requirements.txt
Flask
torch
torchvision
Pillow
backgroundremover
